// src/app/events/event-editor.interface.ts
export interface IEventEditor {
    EventNum?: number;
    PreTitle?: string;
    Title: string;
    StartDate: Date;
    EndDate?: Date;
    Address1: string;
    Address2?: string;
    Address3?: string;
    City?: string;
    Postcode?: string;
    ThumbnailName?: string;
    FlyerName?: string;
    Details?: string;
}